Remove get_itr_snapshot() (#1054)
This commit addresses #1016 . It removes all get_itr_snapshot() from algorithms. This interface was a legacy from rllab to do snapshotting and is deprecated because we now require an algorithm to be pickable/unpickable by rewriting __setstate__() and __getstate__().

Some outdated files in fixtures are not touched. They are used by the sigint test and have diverged from the other parts of the codebase for long.