Disable Gym Time limits by default
A little bit of background about the gym termination signal.

OpenAi gym environments are designed to set the done signal
of environment to True every time step is called, after a
certain number of certain number of environment steps.
This is bad because it changes the MDP of the environment
For gym environments that are time, limit terminated, when
they are terminated, they include an entry in the env_infos
called Timelimit.truncated. This entry is equal to the not
of the termination signal caused by the environment being
in a terminal state (aka the actual/True done signal).
The change I have made to garage/envs/base.py is to set
the done signal to the actual/true done signal whenever
the environment has been timelimit terminated.