Fix nightly tests (#1158)
*  Fix gym viewer windows not closing

The attribute name for gym env entry point changed
from '_entry_point' to 'entry_point' causing viewer
windows to not close, causing memory leak. This commit
fixes that.

* Downgrade and pin atari-py dep to < 0.2.0

Gym 0.13.0 and above depend on atari-py~=0.2.0, but it fails
to load `Defender-*` envs. This commit downgrades to the
the last working version.

* Re-enable nightly tests