Remove atari-py dep;Skip Defender envs in nightly tests (#1194)
* Remove atari-py direct dependency

We pinned atari-py to 0.1.15 because Defender-* envs
refuse to load on atari-py 0.2.x. This conflicts with
gym's dependency on atari-py which breaks pipenv.

* Skip tests for Defender-* envs

gym 0.13.0+ depends on atari-py~=0.2.0 which refuses
to load the Defender-*. Downgrading atari-py brings
additional headaches, so dropping test support for
those envs instead.

If required, Defender.bin in atari-py installation
directory can be replaced by downloading it from
http://www.atarimania.com/rom_collection_archive_atari_2600_roms.html
to make it work.