Fix flaky tests for ray sampler (#1193)
* Fix flaky tests for ray sampler

- Ray sampler would routinely fail Travis CI and cause subsequent
  tests to fail.
- This is because the ray server would be reinited during
  newly added tests in non local ray mode. Local mode in ray makes
  code act serially, which is sufficient for our test suite.
- The newly added tests were not apart of the ray_sampler test class,
  so ray would be inited in standard mode for these tests.
- Standard mode ray consumes more memory than is feasible for our
  travis compute instance.
- The fix was made by adding the tests to the ray_sampler test class,
  as opposed to them being standalone test functions

* Update trpo_ray_sampler.py example

  - Start ray in a reduced state so that example
    uses less resources inside of the CI.
  - Use wrap experiment interface instead of run_experiment
    interface.

* Create pytest fixtures for ray_local_mode

Tensorflow policies can't be pickle and unpickled within
the same session. This means that if we were to run RaySampler
in local mode with tensorflow, the sampler will fail to
construct workers because the workers will try to create
policies whose underlying neural networks can't be copied
into the same session. This could potentially be fixed
if somehow upon unpickling the names of the variables
that make up the policy were to somehow change.

* Convert ray_batch_sampler tests into function

Convert the test_ray_batched_Sampler tests into
functions as opposed to modules, and then convert
the ray_test_fixture into a function wide test fixture
as opposed to a module wide test fixture.