Soft Actor Critic and Applications
This commit adds SAC to garage. The version of
SAC added however is built for collecting batches
of transitions between optimization steps, and not
single transitions, as the original SAC does.

See #807

Furthermore small issues are addressed by this PR such
as #1229 and expanding the robustness for datatypes for
the TanhGaussianMLP Policy.