Restrict test discovery to tests directory (#1256)
Currently pytest starts test discovery in root directory
of garage. This was also causing it to pickup an example
ending with _meta_test as a test. While it is possible to
tell pytest to pick only function of name pattern test_* and
not *_test, restricting the directory is more solid and
avoids unneccasry work.