Make copies of env in worker_factory
When a single object is passed to prepare_worker_messages(), it should
make a list of copies of that object, but not a list of references.

This is a bug that makes test_update_envs_env_update() flaky. Without
this patch, rollouts will be exactly the same, and np.var(mean_rewards)
will be a small number that is close to zero. (~1e-32)