Fix set model param (#1355)
Currently, when setting model parameters, the parameter name has to be an exact
match. This would not work when assigning a model's parameters with a model with
a different name. This PR fixes the bug by only matching the part of the parameter
name without the model name.

(cherry picked from commit aef210dd525d1651fc5966033b91760fe1a126fc)