Extend 2019.10 maintenance window (#1430)
We missed the 2020.02 release, and are on track to release for 2020.06.
Our support policy specifies that releases should get about 8 months of
support, but this also implies an overlap of 4 months to switch from
older to newer releases.

To preserve this overlap, we should support the 2019.10 release for at
least an additional 4 months. Preserving support for the older release
also allows us and users to benefit from a very-stable and yet fairly
full-featured version of garage, giving us more freedom to improve APIs
in 2020.06 and 2020.10.

v2019.10 has proven to be pretty stable and easy to backport into, so I
don't expect this to cause much maintenance burden.