Fix pipenv installation test
The latest release of pipenv doesn't allow installing
multiple extras without removing the lockfile in between,
which was breaking our test.

This test now installs all extras in one step, which is
how it was originally, but got split into 2 because baselines
required TF 1.x while we were on 2. Now baselines is in
the benchmarks package, so we can revert to original way.