Lower MTSAC Reward Threshold
A common algorithm testing pattern in Garage is to run a working
algorithm on a toy environment and verify that it achieves a
return threshold. That return threshold would mean that the
algorithm has learned. The issue with this is that because some
algorithms have high return variance, e.g. mtsac, then finding
a good threshold is tough, and can cause this test to fail, even
if the algorithm hasn't had a performance regression.

The Garage team is replacing this testing style with a continuous
benchmarking system for our supported algorithm.

In the meantime, I'm going to severely reduce the return threshold
to 0 solving #1404, because there isn't another good way to test
this algorithm at the moment. The 0 threshold will verify that the
algorithm is runnable from inside of our CI.