Don't collect more than necessary in RaySampler
There is a poorly-handled case where we have more workers than we need.
At the moment, the fix for this is to only call rollout on workers if
we need samples from those workers. For example, if there are 2
workers, and they are each configured collect 16 samples per call, and
then a call to obtain samples is made asking for only 16 samples, then
only the first sampler worker should be started.

Fixes #1349