Create DATA_PATH dir for docker if not exists
If the host directory, being mounted as a volume in docker using the
-v or --volume option, doesn't exist on the host side, docker creates
that directory with root as user and group. This causes that directory
to be non-writable by garage since the user inside container is not
root.

This commit causes make to create the said directory if it doesn't
exist already