Request review from contributors only once
Mergify's default behavior is to run request_reviews action
on every event like when somebody comments on a PR or approves it etc
This gets annoying since that keeps triggering the github slack
bot and sends message to everybody in garage-contributors.

This commit adds a check for requesting review from
contributors only if it's not already done. This might still
cause it to send 1 extra notification though.

Another option is to check if # of review-requested is
less than, say 3, but those 3 could be all maintainers,
in which case it wouldn't request contributors at all.