Flatten observations if necessary
This commit takes care to convert observations that belong to
spaces that aren't box spaces into observations that would
come from box spaces via akro observation space flattening.

It is really a temporary change. The better solution would be
to modify our GymEnv/GarageEnv and BulletEnv Wrappers to flatten
observations spaces that aren't box spaces.

This is a smaller change, however, the HER replay buffer would
need to be rewritten, which is a much larger change/refactor.

I have refrained from modifying the primitives that belong to
multi task and metalearning algorithms specifically.

Additionally, this change updates ussages of ray to reflect
the update in parameter names from ray 0.8.6->0.8.7