Standardize around "episode" terminology
This PR aims to eliminate all alternative terms for an episode (e.g.
"rollout", "path", "trajectory", etc.) and standardize around the term
"episode" for a single experience in an agent's life.

Notable exception:
I retained the term "path(s)" when referring to the deprecated
dict-based data structure representing an episode, to make refactoring
these code sites easier in the future.