Flatten observations in policies if needed(#1812)
* Standardize around "episode" terminology (#1892)

This PR aims to eliminate all alternative terms for an episode (e.g.
"rollout", "path", "trajectory", etc.) and standardize around the term
"episode" for a single experience in an agent's life.

Notable exception:
I retained the term "path(s)" when referring to the deprecated
dict-based data structure representing an episode, to make refactoring
these code sites easier in the future.

* Flatten observations if necessary

This commit takes care to convert observations that belong to
spaces that aren't box spaces into observations that would
come from box spaces via akro observation space flattening.

It is really a temporary change. The better solution would be
to modify our GymEnv/GarageEnv and BulletEnv Wrappers to flatten
observations spaces that aren't box spaces.

This is a smaller change, however, the HER replay buffer would
need to be rewritten, which is a much larger change/refactor.

I have refrained from modifying the primitives that belong to
multi task and metalearning algorithms specifically.

Additionally, this change updates ussages of ray to reflect
the update in parameter names from ray 0.8.6->0.8.7

Co-authored-by: Ryan Julian <ryanjulian@users.noreply.github.com>