Pin gym version in benchmarks package (#1915)
openai/baselines depends on gym<16.0 while it is pinned
to 0.17.2 in garage. Installing baselines over garage
leads to installation of gym 0.15.7 which isn't compatible
with cloudpickle 1.3. We do this in our Dockerfile.

This is a temporary fix until we can remove gym dependency
in the main garage package

This might break baselines but since we are not actively
using that, we can deal with that later

(cherry picked from commit 6d889604200a2fda58f92a09c809a3a939f8db37)