Backport Various SAC and MTSAC Bug Fixes
Backport #1905, #1975, #1908 to fix problems
with max_eval_path_length being not used by
mtsac and sac, and add checking for incorrect
num_tasks being set in mtsac.

Timelimit.truncated modified only when necessary

This issue occurs when there are multiple garage
envs that are nested or timelimit truncated = False
is included in the environment keys.
Previously, our timelimit
truncated logic was written with the idea in mind
that the key was only added when a time limit
truncation occured. If an environment already
has timelimit truncated = False in its keys
then the previous behavior was to set Done = True
which is the incorrect behavior.

That was causing performance degradation
in MTSAC and MTPPO/TRPO.

Now Done is only true in the normal/trivial case,
never if timelimit truncated is False.