Disable Ray Dashboard
Currently, when we test examples that used the Ray Sampler,
which is every on policy rl algorithm, we shut down the example
after 1 epoch. Because of this, ray.shutdown() is never called
which is necessary for closing ray features such as the ray
dashboard. The dashboard by default tries to attach to the same
port, and so because it hasn't been properly freed after the
example is prematurely shutdown, sometimes the ray sampler
fails to properly init when a new example is started.

To work around this, I have disabled the behavior where the
ray sampler activates its ray dashboard. Users can renable
the dashboard by initing ray inside of their launchers,
like we do in our trpo_swimmer_ray_sampler.py example.

See #1961 for more information