Pin atari-py to last release with packaged ROMs
This is a temporary fix. atari-py releases starting 0.2.7
don't install ROMs by default. While there are instructions
given for how to install ROMs on their github, it is still
incovenient for a garage user and adds an extra step in
installation.

A better fix would be to figure out if it's possible to
install ROMs as a post install step in setup.py